pyproject.toml
[tool.poetry]
name = "lfx-landscape-tools"
version = "20241209"
description = "Collection of tools for automating the data management of a landscape from LFX"
authors = ["John Mertic <jmertic@linuxfoundation.org>"]
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">= 3.12"
requests = "^2.32.3"
pycairo = "^1.27.0"
pygithub = "^2.5.0"
python-slugify = "^8.0.4"
requests-cache = "^1.2.1"
ruamel-yaml = "^0.18.10"
url-normalize = "^1.4.3"
urllib3 = "^2.3.0"
validators = "^0.34.0"
click = "^8.1.8"

[tool.poetry.group.test.dependencies]
coverage = "^7.6.11"
responses = "^0.25.5"

[tool.poetry.group.dev.dependencies]
build = "^1.2.2"

[tool.poetry.scripts]
lfx_landscape = "lfx_landscape_tools.cli:Cli.run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
